Move VirtualMount and EmuConfig to api
