+++ src/main/java/net/clgd/ccemux/emulation/CCEmuX.java
@@ -14,8 +14,9 @@ import dan200.computercraft.core.computer.IComputerEnvironment;
 import dan200.computercraft.core.filesystem.*;
 import lombok.*;
 import lombok.extern.slf4j.Slf4j;
-import net.clgd.ccemux.emulation.filesystem.VirtualDirectory;
-import net.clgd.ccemux.emulation.filesystem.VirtualMount;
+import net.clgd.ccemux.api.emulation.EmuConfig;
+import net.clgd.ccemux.api.emulation.filesystem.VirtualDirectory;
+import net.clgd.ccemux.api.emulation.filesystem.VirtualMount;
 import net.clgd.ccemux.plugins.PluginManager;
 import net.clgd.ccemux.rendering.Renderer;
 import net.clgd.ccemux.rendering.RendererFactory;
