--- src/main/java/net/clgd/ccemux/emulation/filesystem/VirtualMountEntry.java
@@ -1,10 +1,0 @@
-package net.clgd.ccemux.emulation.filesystem;
-
-/**
- * A superinterface for {@link VirtualFile} and {@link VirtualDirectory}, used
- * by {@link VirtualMount}
- * 
- * @author apemanzilla
- *
- */
-public interface VirtualMountEntry { }
