+++ src/main/java/net/clgd/ccemux/plugins/Plugin.java
@@ -11,7 +11,7 @@ import java.util.stream.Collectors;
 
 import lombok.extern.slf4j.Slf4j;
 import net.clgd.ccemux.api.config.Group;
-import net.clgd.ccemux.emulation.EmuConfig;
+import net.clgd.ccemux.api.emulation.EmuConfig;
 import net.clgd.ccemux.plugins.hooks.Hook;
 
 /**
