+++ src/main/java/net/clgd/ccemux/plugins/PluginManager.java
@@ -10,10 +10,10 @@ import lombok.val;
 import lombok.extern.slf4j.Slf4j;
 import net.clgd.ccemux.api.config.ConfigProperty;
 import net.clgd.ccemux.api.config.Group;
+import net.clgd.ccemux.api.emulation.EmuConfig;
+import net.clgd.ccemux.api.emulation.filesystem.VirtualDirectory;
 import net.clgd.ccemux.emulation.CCEmuX;
-import net.clgd.ccemux.emulation.EmuConfig;
 import net.clgd.ccemux.emulation.EmulatedComputer;
-import net.clgd.ccemux.emulation.filesystem.VirtualDirectory;
 import net.clgd.ccemux.plugins.hooks.Closing;
 import net.clgd.ccemux.plugins.hooks.ComputerCreated;
 import net.clgd.ccemux.plugins.hooks.ComputerRemoved;
