+++ src/main/java/net/clgd/ccemux/plugins/builtin/AWTPlugin.java
@@ -11,7 +11,7 @@ import com.google.auto.service.AutoService;
 
 import net.clgd.ccemux.api.config.ConfigProperty;
 import net.clgd.ccemux.api.config.Group;
-import net.clgd.ccemux.emulation.EmuConfig;
+import net.clgd.ccemux.api.emulation.EmuConfig;
 import net.clgd.ccemux.emulation.EmulatedComputer;
 import net.clgd.ccemux.plugins.Plugin;
 import net.clgd.ccemux.rendering.Renderer;
