+++ src/main/java/net/clgd/ccemux/plugins/builtin/CCEmuXAPI.java
@@ -16,11 +16,11 @@ import dan200.computercraft.api.lua.LuaException;
 import dan200.computercraft.core.apis.ArgumentHelper;
 import dan200.computercraft.core.apis.ILuaAPI;
 import lombok.extern.slf4j.Slf4j;
+import net.clgd.ccemux.api.emulation.EmuConfig;
+import net.clgd.ccemux.api.emulation.filesystem.VirtualFile;
+import net.clgd.ccemux.api.emulation.filesystem.VirtualDirectory.Builder;
 import net.clgd.ccemux.emulation.CCEmuX;
-import net.clgd.ccemux.emulation.EmuConfig;
 import net.clgd.ccemux.emulation.EmulatedComputer;
-import net.clgd.ccemux.emulation.filesystem.VirtualDirectory.Builder;
-import net.clgd.ccemux.emulation.filesystem.VirtualFile;
 import net.clgd.ccemux.plugins.Plugin;
 import net.clgd.ccemux.plugins.hooks.ComputerCreated;
 import net.clgd.ccemux.plugins.hooks.CreatingROM;
