+++ src/main/java/net/clgd/ccemux/plugins/builtin/TRoRPlugin.java
@@ -7,7 +7,7 @@ import java.util.Set;
 
 import com.google.auto.service.AutoService;
 
-import net.clgd.ccemux.emulation.EmuConfig;
+import net.clgd.ccemux.api.emulation.EmuConfig;
 import net.clgd.ccemux.plugins.Plugin;
 import net.clgd.ccemux.rendering.RendererFactory;
 import net.clgd.ccemux.rendering.tror.TRoRRenderer;
