+++ src/main/java/net/clgd/ccemux/rendering/awt/AWTRenderer.java
@@ -46,8 +46,8 @@ import lombok.Getter;
 import lombok.val;
 import lombok.extern.slf4j.Slf4j;
 import net.clgd.ccemux.api.Utils;
+import net.clgd.ccemux.api.emulation.EmuConfig;
 import net.clgd.ccemux.emulation.CCEmuX;
-import net.clgd.ccemux.emulation.EmuConfig;
 import net.clgd.ccemux.emulation.EmulatedComputer;
 import net.clgd.ccemux.plugins.builtin.AWTPlugin.AWTConfig;
 import net.clgd.ccemux.rendering.Renderer;
